# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KhelSetu API
  version: 1.0.0
  description: >
    # KhelSetu API Documentation


    ## Platform Overview


    KhelSetu is a multi-tenant sports tournament management platform with **two
    distinct portals**:


    ### 🌐 Public Portal (No Login Required)

    The public portal is for **viewers and fans** who want to watch live
    matches, browse tournaments, and view standings. These endpoints are
    **completely open** — no authentication needed.


    - **Live match scores** — Real-time scores for any match

    - **Tournament browsing** — View all public tournaments

    - **Standings & fixtures** — League tables and match schedules

    - **Team & player profiles** — Roster and player information

    - **Global search** — Search across tournaments, teams, and players


    ### 🔒 SaaS Dashboard (Login Required)

    The dashboard is for **organization members** (admins, scorers, coaches) who
    manage tournaments, teams, players, and live scoring. **All dashboard
    endpoints require JWT authentication** and **organization membership**.


    - **Organization management** — Create/manage organizations, invite members

    - **Tournament management** — Create tournaments, generate fixtures

    - **Team & player management** — Roster management, player profiles

    - **Live scoring** — Real-time event-driven scoring (Football, Cricket,
    Basketball)

    - **Standings** — Auto-calculated league tables

...

...

...


...


...

...
    refresh token (7d)
...
paths:
...
  /api/visualization/formations:
    post:
      summary: Create formation
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Formation created
  /api/visualization/formations/match/{matchId}/team/{teamId}:
    get:
      summary: Get formation for match
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Formation data
  /api/visualization/formations/{id}:
    put:
      summary: Update formation
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Formation updated
  /api/visualization/annotations:
    post:
      summary: Create annotation
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Annotation created
  /api/visualization/annotations/match/{matchId}:
    get:
      summary: Get match annotations
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of annotations
  /api/visualization/annotations/{id}:
    delete:
      summary: Delete annotation
      tags:
        - Visualization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Deleted
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token. Get it from POST /api/auth/login. Pass as:
        Authorization: Bearer <token>"
...
security:
  - bearerAuth: []
...